# api/openapi.yaml
...
components:
  schemas:
    ErrorResponse:
      properties:
        error:
          type: string
      type: object
    MainUsage:
      properties:
        input_tokens:
          type: integer
        output_tokens:
          type: integer
      type: object
...
    RawTurnHeaderItem:
      properties:
        agent_name:
          type: string
        id:
          type: integer
        meta:
          type: object
        provider:
          type: string
        received_at:
          type: string
        request_bytes:
          type: integer
        request_id:
          type: string
        response_bytes:
          type: integer
        source:
          type: string
      type: object
    RawTurnListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/RawTurnHeaderItem'
          type: array
      type: object
...
    SpanItem:
      properties:
        call_kind:
          description: Deriver-written taxonomy, promoted from the old metadata grab-bag.
          type: string
        duration_ns:
          type: integer
        input:
          description: |-
            Input/Output are content-block arrays (llm.ContentBlock), uniform for
            every kind (tool spans included — no unwrapping). Pinned to [] when
            empty. swaggertype keeps the wire an array-of-object rather than the
            json.RawMessage byte array swag infers.
          items:
            type: object
          type: array
        kind:
          type: string
        model:
          type: string
        name:
          type: string
        output:
          items:
            type: object
          type: array
        parent_span_id:
          type: string
        payload:
          description: |-
            Payload marks a preview-truncated span so the console drills in for
            the full payload; absent in full mode.
          type: string
        raw_turn_id:
          type: integer
        seq:
          description: |-
            Seq is the span's presentation ordinal within its trace; spans
            arrive sorted by it (started_at ties inside one llm call — parallel
            tool batches share an instant).
          type: integer
        span_id:
          type: string
        started_at:
          type: string
        status:
          type: string
        stop_reason:
          type: string
        thread_id:
          type: string
        trace_id:
          type: string
        usage:
          description: |-
            Usage (was `metrics`) is an llm.Usage object on the wire — {}-pinned
            for usage-less spans.
          type: object
        verdict:
          description: |-
            Verdict is the typed security-monitor disposition (null off
            permission-check spans), deriver-written. It is a Verdict object or
            null on the wire; typed as `object` (not the json.RawMessage byte
            array swag would otherwise emit) via swaggertype.
          nullable: true
          type: object
      type: object
    SpanLinkItem:
      properties:
        from_io:
          type: string
        from_span_id:
          type: string
        from_trace_id:
          type: string
        kind:
          type: string
        to_io:
          type: string
        to_span_id:
          type: string
        to_trace_id:
          type: string
      type: object
...
    TraceDetail:
      properties:
        links:
          items:
            $ref: '#/components/schemas/SpanLinkItem'
          type: array
        schema:
          type: string
        spans:
          items:
            $ref: '#/components/schemas/SpanItem'
          type: array
        trace:
          $ref: '#/components/schemas/TraceItem'
      type: object
    TraceItem:
      properties:
        duration_ns:
          type: integer
        ended_at:
          type: string
        main_usage:
          $ref: '#/components/schemas/MainUsage'
        response_preview:
          description: |-
            ResponsePreview is the derive-time fold of the closing
            conversation-spine llm call's text output — the answer line for
            collapsed turn cards, so summary consumers never need spans.
          type: string
        source:
          description: |-
            Source is the capture origin of the turn's rows ("wire" |
            "transcript"), promoted from raw_turns.source. Per-trace, so a
            session can mix live wire capture and transcript backfill. Today
            every trace is "wire": transcripts only reconcile fork/parent edges
            during derivation, they never form a trace on their own. "transcript"
            becomes real when a session is reconstructed purely from a transcript
            file with no proxy capture (an OSS backfill path).
          type: string
        span_count:
          type: integer
        started_at:
          type: string
        status:
          type: string
        synthetic:
          description: |-
            Synthetic is a typed deriver signal ("post-compaction" for a
            compaction continuation, "shadow-opener" for a shadow-only opener),
            promoted out of the old metadata grab-bag. Absent for genuine
            prompt-opened turns.
          type: string
        trace_id:
          type: string
        usage:
          allOf:
          - $ref: '#/components/schemas/TraceUsage'
          description: |-
            Usage is the trace's total token/cost spend over ALL llm spans,
            shadow calls included; MainUsage is the task slice — the main agent
            and its subagents (every call_kind=main span, across threads). The
            difference (Usage − MainUsage) is the harness's shadow spend
            (permission checks, title-gen, web summaries) on the turn.
        user_prompt:
          description: |-
            UserPrompt is served explicitly (not omitempty): a synthetic opener
            has an empty prompt, and dropping the key turns the empty string
            into `undefined` on the wire, which breaks consumers that expect a
            string (e.g. the console's stripHarnessTags). Empty means synthetic.
          type: string
      type: object
    TraceListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/TraceItem'
          type: array
        schema:
          type: string
      type: object
    TraceUsage:
      properties:
        cache_creation_tokens:
          type: integer
        cache_read_tokens:
          type: integer
        cost_usd:
          type: number
        input_tokens:
          type: integer
        output_tokens:
          type: integer
      type: object
...
info:
  contact: {}
  description: |-
    HTTP API for inspecting, querying, and searching stored Tapes sessions.

    The REST surface exposes health checks, session listing and retrieval, derived session summaries, aggregate stats, semantic search, skill authoring and publishing, operator maintenance endpoints, and a streamable MCP endpoint.

    Not covered here: the ingest write surface, which is a separate server with its own contract (see ingest/openapi.yaml).
  title: Tapes API
  version: "1.0"
openapi: 3.0.3
paths:
...
  /v1/sessions/{id}/raw_turns:
    get:
      description: 'The raw layer''s wire log: one row per captured call or transcript
        push, identity and sizes only. `source` distinguishes what crossed the wire
        from what the harness pushed as its own account.'
      operationId: listRawTurns
      parameters:
      - description: Session id (UUID)
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RawTurnListResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
      summary: List a session's raw capture log (operator)
      tags:
      - sessions
...
  /v1/traces:
    get:
      description: Returns turn headers for a session — no span payloads. Fetch GET
        /v1/traces/{trace_id} per turn for spans and links.
      operationId: listTraces
      parameters:
      - description: Session id (UUID)
        in: query
        name: session_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceListResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
      summary: List a session's traces (summaries)
      tags:
      - traces
  /v1/traces/{trace_id}:
    get:
      description: 'Returns one user-visible turn: its spans nested by parent_span_id
        and its dataflow links (links touching other traces included).'
      operationId: getTrace
      parameters:
      - description: Trace id
        in: path
        name: trace_id
        required: true
        schema:
          type: string
      - description: 'Span payload mode: full (default) or preview (strings truncated;
          fetch the span endpoint for full payloads)'
        in: query
        name: payload
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceDetail'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
      summary: Get one trace with spans and links
      tags:
      - traces
  /v1/traces/{trace_id}/spans/{span_id}:
    get:
      description: 'The payload drill-in: one span''s complete input/output content.'
      operationId: getSpan
      parameters:
      - description: Trace id
        in: path
        name: trace_id
        required: true
        schema:
          type: string
      - description: Span id
        in: path
        name: span_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpanItem'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
      summary: Get one span with full payloads
      tags:
      - traces